+++ transform/transformTokens.js
@@ -1,8 +1,7 @@
 const StyleDictionary = require("style-dictionary");
-const { getConfig } = require("./config");
-const { javascriptEsm } = require("./src/format/javascriptEsm");
-const { webShadows } = require("./src/transform/webShadows");
-const { transform } = require("typescript");
+const { getConfig } = require("./config.js");
+const { javascriptEsm } = require("./format/javascriptEsm");
+const { webShadows } = require("./transform/webShadows");
 
 function log(x) {
   console.log(`🤖 ${x}`);
